pushed common code down to AutoBase.java
