+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmode/auto/Auto2_BlueFieldRight.java
@@ -85,52 +85,21 @@ import org.openftc.easyopencv.OpenCvWebcam;
 
 @Autonomous(name="BlueFieldRight", group="OpMode")
 //@Disabled
-public class Auto2_BlueFieldRight extends OpMode {
-    // Declare OpMode members for each of the 4 motors.
-    private ElapsedTime runtime = new ElapsedTime();
-    private DcMotor leftFrontDrive = null;
-    private DcMotor leftBackDrive = null;
-    private DcMotor rightFrontDrive = null;
-    private DcMotor rightBackDrive = null;
+public class Auto2_BlueFieldRight extends AutoBase {
 
-    private DcMotor leftLinearSlide = null;
-    private DcMotor rightLinearSlide = null;
-
-    Servo leftClaw;
-    Servo rightClaw;
-    Servo wrist;
-    Servo conveyor;
-    private IMU imu;
-
-    static final int low_linearslide_ticks = 200; // Low position for the linear slides
-    static final int bottom_linearslide_ticks = 0; // Bottom position for the linear slides
-    static final int STREAM_WIDTH = 1280; // modify for your camera
-    static final int STREAM_HEIGHT = 960; // modify for your camera
-    OpenCvWebcam webcam;
-    GripPipelineWhitePixelRGBT1 pipeline;
-
-    Movement moveTo;
-
-    IntakeMovement intake;
-
-    LinearSlideMovement linearSlideMove;
-
-    private GamePieceLocation gamepieceLocation = GamePieceLocation.LEFT;
-
-    int state;
-
-    double rightCount = 0;
-    double centerCount = 0;
-    double leftCount = 0;
-
+    @Override
+    public void init() {
+        super.init();
+        gamepieceLocation = GamePieceLocation.LEFT; // this is the position that we can't see
+    }
     @Override
     public void init_loop(){
         state = 0;
-        GamepiecePositionFinder gamepiecePOS = new GamepiecePositionFinder(pipeline.avgContourCoord(), GamePieceLocation.RIGHT);
+        GamepiecePositionFinder gamePiecePOS = new GamepiecePositionFinder(pipeline.avgContourCoord(), GamePieceLocation.RIGHT);
         Point avgLoc = pipeline.avgContourCoord();
-        if (gamepiecePOS.getPOS() == GamePieceLocation.RIGHT){
+        if (gamePiecePOS.getPOS() == GamePieceLocation.RIGHT){
             rightCount += 1;
-        } else if (gamepiecePOS.getPOS() == GamePieceLocation.CENTER){
+        } else if (gamePiecePOS.getPOS() == GamePieceLocation.CENTER){
             centerCount += 1;
         } else {
             leftCount += 1;
@@ -156,103 +125,6 @@ public class Auto2_BlueFieldRight extends OpMode {
         telemetry.addData("state", state);
         telemetry.update();
     }
-    @Override
-    public void init() {
-
-        int cameraMonitorViewId = hardwareMap.appContext.getResources().getIdentifier("cameraMonitorViewId", "id", hardwareMap.appContext.getPackageName());
-        WebcamName webcamName = null;
-        webcamName = hardwareMap.get(WebcamName.class, "gge_cam"); // put your camera's name here
-        webcam = OpenCvCameraFactory.getInstance().createWebcam(webcamName, cameraMonitorViewId);
-        //pipeline = new GripPipelineRedGamepieceRGB();
-        pipeline = new GripPipelineWhitePixelRGBT1();
-        webcam.setPipeline(pipeline);
-
-        webcam.openCameraDeviceAsync(new OpenCvCamera.AsyncCameraOpenListener()
-        {
-            @Override
-            public void onOpened()
-            {
-                webcam.startStreaming(STREAM_WIDTH, STREAM_HEIGHT, OpenCvCameraRotation.UPRIGHT);
-            }
-
-            @Override
-            public void onError(int errorCode) {
-                telemetry.addData("Camera Failed","");
-                telemetry.update();
-            }
-        });
-
-        // Initialize the hardware variables. Note that the strings used here must correspond
-        // to the names assigned during the robot configuration step on the DS or RC devices.
-        leftFrontDrive = hardwareMap.get(DcMotor.class, "left_front_drive");
-        leftBackDrive = hardwareMap.get(DcMotor.class, "left_back_drive");
-        rightFrontDrive = hardwareMap.get(DcMotor.class, "right_front_drive");
-        rightBackDrive = hardwareMap.get(DcMotor.class, "right_back_drive");
-
-        leftLinearSlide = hardwareMap.get(DcMotor.class, "left_linear_slide");
-        rightLinearSlide = hardwareMap.get(DcMotor.class, "right_linear_slide");
-
-        leftClaw = hardwareMap.get(Servo.class, "left_claw");
-        rightClaw = hardwareMap.get(Servo.class, "right_claw");
-        wrist = hardwareMap.get(Servo.class, "wrist");
-
-        conveyor = hardwareMap.get(Servo.class, "conveyor");
-
-        imu = hardwareMap.get(IMU.class, "imu");
-
-        double DirectionNow = imu.getRobotYawPitchRollAngles().getYaw(AngleUnit.DEGREES);
-
-        // Adding in PIDF Config values learned from previous testing
-        // These may need to be tuned anytime the motor weights or config changes.
-        // Set PIDF values thinking of ...
-        // ...P as primary force (set second)
-        // ...I as smoothing (set last)
-        // ...D as deceleration (set third)
-        // ...F as holding / static force (set first)
-        // For Mecanum drive, 8, 0, 0.5, 5 works well on Tiny
-        // ... and 7, 0.2, 0.1, 8 works on Rosie (heavier bot)
-        ((DcMotorEx) leftFrontDrive).setVelocityPIDFCoefficients(8, 0.1, 0.2, 8);
-        ((DcMotorEx) leftBackDrive).setVelocityPIDFCoefficients(8, 0.1, 0.2, 8);
-        ((DcMotorEx) rightFrontDrive).setVelocityPIDFCoefficients(8, 0.1, 0.2, 8);
-        ((DcMotorEx) rightBackDrive).setVelocityPIDFCoefficients(8, 0.1, 0.2, 8);
-        // For Lift, PIDF values set to reduce jitter on high lift
-        ((DcMotorEx) leftLinearSlide).setVelocityPIDFCoefficients(8, 0.75, 0, 4);
-        ((DcMotorEx) rightLinearSlide).setVelocityPIDFCoefficients(8, 0.75, 0, 4);
-
-        // ########################################################################################
-        // !!!            IMPORTANT Drive Information. Test your motor directions.            !!!!!
-        // ########################################################################################
-        // Most robots need the motors on one side to be reversed to drive forward.
-        // The motor reversals shown here are for a "direct drive" robot (the wheels turn the same direction as the motor shaft)
-        // If your robot has additional gear reductions or uses a right-angled drive, it's important to ensure
-        // that your motors are turning in the correct direction.  So, start out with the reversals here, BUT
-        // when you first test your robot, push the left joystick forward and observe the direction the wheels turn.
-        // Reverse the direction (flip FORWARD <-> REVERSE ) of any wheel that runs backward
-        // Keep testing until ALL the wheels move the robot forward when you push the left joystick forward.
-        leftFrontDrive.setDirection(DcMotor.Direction.FORWARD); // Somehow this is reversed from the TeleOp Gge program.
-        leftBackDrive.setDirection(DcMotor.Direction.FORWARD);
-        rightFrontDrive.setDirection(DcMotor.Direction.REVERSE);
-        rightBackDrive.setDirection(DcMotor.Direction.REVERSE);
-
-        leftLinearSlide.setDirection(DcMotorSimple.Direction.FORWARD);
-        rightLinearSlide.setDirection(DcMotorSimple.Direction.REVERSE);
-
-        intake = new IntakeMovement(rightClaw, leftClaw, wrist, telemetry);
-        moveTo = new Movement(leftFrontDrive, rightFrontDrive, leftBackDrive, rightBackDrive, imu, telemetry);
-        linearSlideMove = new LinearSlideMovement(leftLinearSlide, rightLinearSlide, intake);
-
-        state = 0;
-        //drive speed limiter
-        //double powerFactor = 0.25;
-
-        // Wait for the game to start (driver presses PLAY)
-        telemetry.addData("Status", "Initialized");
-
-        telemetry.update();
-
-        runtime.reset();
-    }
-        // run until the end of the match (driver presses STOP)
 
     @Override
     public void loop(){
@@ -318,14 +190,5 @@ public class Auto2_BlueFieldRight extends OpMode {
         }
 
         // Show the elapsed game time and wheel power.
-        telemetry.addData("Status", "Run Time: " + runtime.toString());
-        telemetry.addData("Direction Now", JavaUtil.formatNumber(DirectionNow, 2));
-        telemetry.addData("Target Position", leftFrontDrive.getTargetPosition());
-        telemetry.addData("Left Front Pos", leftFrontDrive.getCurrentPosition());
-        telemetry.addData("Right Front Pos", rightFrontDrive.getCurrentPosition());
-        telemetry.addData("Left Back Pos", leftBackDrive.getCurrentPosition());
-        telemetry.addData("Right Back Pos", rightBackDrive.getCurrentPosition());
-        telemetry.addData("state", state);
-        telemetry.addData("location", gamepieceLocation);
-        telemetry.update();
+        displayTelemetry(DirectionNow);
     }}
